requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2

# Deep Learning
tensorflow==2.13.0
torch==2.0.1

# API
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3

# Data validation
great-expectations==0.17.12

# MLOps tools
mlflow==2.7.1
dvc==3.31.1
wandb==0.15.8

# Monitoring and logging
prometheus-client==0.18.0
python-json-logger==2.0.7

# Database
sqlalchemy==2.0.21
psycopg2-binary==2.9.7

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# Code quality
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.5.1
pylint==2.17.7

# Development
jupyter==1.0.0
jupyterlab==4.0.4
ipython==8.14.0
python-dotenv==1.0.0

# Utilities
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
requests==2.31.0
joblib==1.3.2
